requirements.txt
scrapy
elasticsearch
scrapy_proxies
python-dateutil
beautifulsoup4
scrapy-redis
pysummarization